server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
sqlalchemy-utils==0.41.1
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi-cors==0.0.6
email-validator==2.1.0
